requirements.txt
asgiref==3.10.0
Django==5.2.7
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
drf-spectacular==0.27.2
Pillow==11.0.0
Faker==32.0.0
django-filter==24.3
